+++ admin/menu/rumah_kategori/index.php
@@ -25,10 +25,14 @@
 
 	/* for delete */
 	if (isset($_GET["delete"])) {
-		$idDelete = $_GET["delete"];
-		$hapus = $model->delete($idDelete);
-		if ($hapus) {
+		if ($admin == null) {
 			echo "<script> document.location.href = '".$backRedirect."'; </script>";
+		} else {
+			$idDelete = $_GET["delete"];
+			$hapus = $model->delete($idDelete);
+			if ($hapus) {
+				echo "<script> document.location.href = '".$backRedirect."'; </script>";
+			}
 		}
 	}
 ?>
